+++ packages/cdn/lib/ethers.ts
@@ -1,2 +1,2 @@
 import '@rerock/polyfills'
-export * as Web3modal from '@rerock/ethers'
+export * as Web3modal from '@rerock/appkit-ethers'