+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @web3modal/ethers
 
+## 5.1.9
+
+### Patch Changes
+
+- Adds CoinbaseConnector to Solana package
+
+- Updated dependencies []:
+  - @web3modal/scaffold-utils@5.1.9
+  - @web3modal/polyfills@5.1.9
+  - @web3modal/common@5.1.9
+  - @web3modal/wallet@5.1.9
+  - @web3modal/base@5.1.9
+  - @web3modal/siwe@5.1.9
+
 ## 5.1.8
 
 ### Patch Changes